# PERF: fastpaths in is_foo_dtype checks