@jorisvandenbossche 2020-04-10
Related to the discussion we had about this on the chat, to give an example of the strategy that I tried to explain:

The current `is_categorical_dtype` is defined as:

```python
def is_categorical_dtype(arr_or_dtype) -> bool:
    if arr_or_dtype is None:
        return False
    return CategoricalDtype.is_dtype(arr_or_dtype)
```

where `CategoricalDtype.is_dtype()` is the catch-all / check-all slow check.

So for this case, we can put the fast check as in this PR first:

```python
def is_categorical_dtype(arr_or_dtype) -> bool:
    if isinstance(dtype, ExtensionDtype):
        # fast check for extension dtype
        return dtype.name == "category"
    elif isinstance(dtype, np.dtype):
        # fast check for numpy dtype (always False)
        return False
    elif arr_or_dtype is None:
        return False
    # keep the slow check if the fast ones didn't return
    return CategoricalDtype.is_dtype(arr_or_dtype)
```

And in that way, we don't have to change every occurrence of `is_categorical_dtype` internally to the fast function.

It might be that categorical is a simple example, but I would think that a similar pattern can be followed for the others as well.