@jbrockmendel 2020-04-10
> Related to the discussion we had about this on the chat, to give an example of the strategy that I tried to explain:

Also as discussed on the chat, the advantage of the dtype-only versions is that we dont risk silently going down the slow path.