@jbrockmendel 2020-04-16
I'm happy to update the existing functions to be faster, but we should also have the strict versions internally to make sure we actually _use_ the fastpath