@jbrockmendel 2020-04-16
Actually, adding the fastpath to the existing checks will actually slow them down slightly in cases where we're not passing a dtype obj