@TomAugspurger 2020-04-16
I think I'm OK with slowing down the non-dtype case.

I'm also OK with deprecating non-dtype entirely, but that's a larger task that can be saved for later. Short-term, I think we can rely on reviewers to catch places where we're passing non-dtype's to `is_foo_dtype`.