@jbrockmendel 2020-04-16
One more doomed pitch for the strict versions: they're really nice dependency-structure-wise.  ATM dtypes.common depends on dtypes.dtypes, which has runtime imports and depends on a bunch of the code.  With just the strict versions, we can get dtypes.common (which we import from basically everywhere) to depend only on things "above" it in the dependency structure.

I know others dont care about the dependency structure as much as I do, will update.